MotionSearcher.hh SetRegion2D.hh main.c newdenoise.cc newdenoise.hh
Implemented the first, easy version of parallelism.  "-p 0" produces
the original behavior.  "-p 1" moves the reading/writing of video
frames to separate threads.  "-p 2" moves the denoising of color to a
separate thread.  Some work was needed to separate their memory spaces
better, using private allocators for the set-region objects.

This does pretty well, but since color generally takes far less time to
denoise than intensity, it still doesn't fill up 2 processors.  The
next step is to rewrite denoising so that 2 search-windows can process
different slices of the same frame at the same time, but that is a BIG
project.

Also, the "skip" option was removed.


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code@2613 3be20b7b-d6de-45d7-9b51-1db703c412de