For some time now there's been the ability to selectively disable accelerated functions (mjpeg_simd_helper gives a list of the functions for which this can be done).  Alas, that only worked for the X86 (MMX, SSE) functions.  The changes being checked in now enable the same capability for the Altivec routines
This is being done to work around an apparent limitation in the sad_00 routine.
yuvdeinterlace wants to use sad_00 with data aligned on 8 byte boundaries.
This works for MMX, SSE but not Altivec.  So now with the simd_disable()
implemented it will be possible to disable sad_00_altivec within yuvdeinterlace.


git-svn-id: https://svn.code.sf.net/p/mjpeg/Code@3142 3be20b7b-d6de-45d7-9b51-1db703c412de